chore: 컴포넌트 디렉토리 구조 기능 단위로 변경
