+++ src/components/auth/AuthWrapper/index.tsx
@@ -1,8 +1,13 @@
 import { Link as RouterLink } from "react-router-dom";
-import { Button, Link as ChakraLink, Flex, Text } from "@chakra-ui/react";
+import { Box, Button, Link as ChakraLink, Flex, Text } from "@chakra-ui/react";
 import kakao_icon from "../../../assets/icon/kakao_icon.svg";
 import naver_icon from "../../../assets/icon/naver_icon.svg";
 import google_icon from "../../../assets/icon/google_icon.png";
+import { ToppestText } from "../ToppestText";
+import { MoveLeft } from "lucide-react";
+import { MouseEventHandler } from 'react';
+
+
 export interface AuthBottomWrapperProps {
     linkText?: string;
     linkText2?: string;
@@ -71,3 +76,39 @@ export const AuthBottomWrapper = ({linkText, linkText2, linkTo}) => {
   );
 };
 
+
+
+export interface AuthTopWrapperProps {
+  text?: string;
+  color?: string;
+  onClick?: MouseEventHandler<HTMLButtonElement>;
+}
+
+const goBack = () => {
+  window.history.go(-1); // 뒤로 가기
+};
+
+
+export const AuthTopWrapper = ({ text, color }: AuthTopWrapperProps) => {
+  return (
+    <Flex w="100%" mt="40px" position="relative" >
+        <Box>
+        <Button m="0" 
+        p="0" 
+        bg="none" 
+        _hover= {{bg:"none"}} 
+        onClick={goBack}
+        width="28px" // 아이콘의 너비와 동일하게 설정
+        height="28px" // 아이콘의 높이와 동일하게 설정
+        display="flex" // 아이콘을 버튼 안에 중앙 정렬
+        alignItems="center"
+        justifyContent="center">
+      <MoveLeft size={28} color={color} />
+      </Button>
+      </Box>
+      <Box position="absolute" right="calc((100% - 60px) / 2)" >
+      <ToppestText text={text} color={color} />
+      </Box>
+    </Flex>
+  );
+};
